try to parallelize UI tests; split them into multiple classes
