+++ Tests/UITests/TestApp.xctestplan
@@ -26,6 +26,7 @@
   },
   "testTargets" : [
     {
+      "parallelizable" : true,
       "target" : {
         "containerPath" : "container:UITests.xcodeproj",
         "identifier" : "2F6D13AB28F5F386007C25D6",
